# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: FinIntel Core REST API
  description: Canonical REST API contract between Next.js Web frontend client and Go Core API.
  version: 1.0.0
  license:
    name: Proprietary
    identifier: Proprietary
servers:
  - url: http://localhost:8080
    description: Local Development Server
security:
  - BearerAuth: []
paths:
...
  /api/v1/organizations/{organizationId}/members/{memberId}/role:
    put:
      summary: Update Member Role
      description: Updates the canonical role of an existing organization member.
      operationId: updateMemberRole
      security:
        - BearerAuth: []
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: memberId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  type: string
                  enum: [OWNER, ADMINISTRATOR, ACCOUNTANT, ANALYST, AUDITOR_VIEWER]
      responses:
        '200':
          description: Role updated successfully.
        '400':
          description: Invalid input or last owner demotion prohibited.
        '403':
          description: Permission denied.
  /api/v1/organizations/{organizationId}/members/{memberId}:
    delete:
      summary: Remove Member
      description: Removes a member from an organization.
      operationId: removeMember
      security:
        - BearerAuth: []
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: memberId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Member removed successfully.
        '400':
          description: Cannot remove last owner.
        '403':
          description: Permission denied.
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...